for #138
first ideas for the new language system in the fastboot-assistant

+++ Anwendung/header/flash/header_flash_heimdall.c
@@ -26,28 +26,28 @@
 #include "program_functions.h"
 #include "language_check.h"
 
-// Global vars
+// global vars
 GtkWidget *spinner_heimdall;
 GtkWidget *spinner_window_heimdall;
 
-// Function to run the command
+// function to run the command
 void *run_heimdall_command(void *command) 
 {
     char *function_command = (char *)command;
 
-    // Run command
+    // run command
     LOG_INFO("Run: %s", function_command);
     system(function_command);
 
-    // Close spinner and window
+    // close spinner and window
     gtk_spinner_stop(GTK_SPINNER(spinner_heimdall));
     gtk_window_destroy(GTK_WINDOW(spinner_window_heimdall));
 
-    free(function_command);  // Free memory
+    free(function_command);  // free memory
     return NULL;
 }
 
-// Function to flash with heimdall
+// function to flash with heimdall
 void flash_heimdall(GtkWidget *widget, GtkWindow *parent_window, const char *partition, const char *image_name) 
 {
     if (!parent_window) parent_window = main_window;
@@ -70,41 +70,49 @@ void flash_heimdall(GtkWidget *widget, GtkWindow *parent_window, const char *par
     }
     
 	snprintf(image_info, sizeof(image_info), "%s/%s", image_path, image_name);
-    // Check if the image exists
+    // check if the image exists
     if (access(image_info, F_OK) == -1) 
     {
         char error_message[4096];
-        snprintf(error_message, sizeof(error_message), g_strcmp0(language, "de") == 0 ? "Image nicht '%s' gefunden.\n" : "Image file '%s' not found.\n", image_info);
+        snprintf(error_message, sizeof(error_message), _("Image file '%s' not found.\n"), image_info);
         show_error_message(GTK_WIDGET(parent_window), error_message);
         return;
     }
-
-    // Create the command
-    char *function_command = malloc(4096);  // Erhöhen des Puffers für längere Kommandos
+    
+    // prevention of crashes
+    if (!is_android_device_connected_fastboot()) 
+    {      
+        const char *error_message = _("No device detected.");
+        show_error_message(GTK_WIDGET(main_window), error_message);
+        return;
+    }
+    
+    // create the command
+    char *function_command = malloc(4096); // bigger buffer for bigger commands
     snprintf(function_command, 4096, "heimdall flash --%s %s --no-reboot", partition, image_info);
 
-    // Create new window
+    // create new window
     spinner_window_heimdall = gtk_window_new();
     gtk_window_set_title(GTK_WINDOW(spinner_window_heimdall), " ");
     gtk_window_set_default_size(GTK_WINDOW(spinner_window_heimdall), 200, 200);
 
-    // Create the spinner
+    // create the spinner
     spinner_heimdall = gtk_spinner_new();
     GtkWidget *vbox = gtk_box_new(GTK_ORIENTATION_VERTICAL, 5);
     gtk_box_append(GTK_BOX(vbox), spinner_heimdall);
     gtk_widget_set_size_request(spinner_heimdall, 150, 150);
     gtk_window_set_child(GTK_WINDOW(spinner_window_heimdall), vbox);
 
-    // Start the spinner
+    // start the spinner
     gtk_spinner_start(GTK_SPINNER(spinner_heimdall));
 
-    // Show new window (using gtk_widget_set_visible instead of deprecated gtk_widget_show)
+    // show new window (using gtk_widget_set_visible instead of deprecated gtk_widget_show)
     gtk_widget_set_visible(spinner_window_heimdall, TRUE);
 
-    // Run the command in a new thread
+    // run the command in a new thread
     pthread_t thread;
     pthread_create(&thread, NULL, run_heimdall_command, function_command);
-    pthread_detach(thread);  // Thread in the background
+    pthread_detach(thread);  // thread in the background
     
     if (image_path != NULL) 
 	{
